wip not working but basic code completed

+++ app/src/main/java/rak/pixellwp/drawing/CirclePoint.kt
@@ -1,3 +1,3 @@
 package rak.pixellwp.drawing
 
-data class CirclePoint(val text: String, val x: Int, val y: Int)
+data class CirclePoint(val text: String, val x: Float, val y: Float)
